requirements.txt
azure-ai-projects==2.0.1
azure.ai.agents==1.1.0
azure-identity==1.25.3
python-dotenv==1.2.2